[Feature] Add a switch for the background layout experiment
Add an experimental switch for the LynxTemplateRender#mEnableLayoutOnBackgroundThread. When this value is false, the layout process runs on the UI thread or a background thread depending on the different thread strategy. When this value is true, regardless of the user's thread strategy, the layout process will run on a background thread. With this switch, we can conduct an experiment to verify whether the layout process runs on a background thread by default can bring benefits.

issue: m-5925896507